Closes #2203: Move setChannelMarginByPosition to ChannelTile.companion
