+++ app/src/main/java/org/mozilla/tv/firefox/navigationoverlay/channels/ChannelLayoutManager.kt
@@ -12,7 +12,6 @@ import androidx.recyclerview.widget.LinearSmoothScroller
 import androidx.recyclerview.widget.RecyclerView
 import org.mozilla.tv.firefox.R
 import org.mozilla.tv.firefox.ext.getDimenPixelSize
-import org.mozilla.tv.firefox.ext.setHorizontalMargins
 
 const val MILLISECONDS_PER_INCH = 50f // For smooth scrolling speed
 
@@ -49,7 +48,7 @@ class ChannelLayoutManager(
             // For carousel scrolling margin updates. Precondition here is that the channel
             // RecyclerView has gained focus. Mainly for handling remove tiles (the adjacent tile
             // gains focus upon removal)
-            it.setHorizontalMargins(context, pos, itemCount)
+            ChannelTile.setChannelMarginByPosition(it, context, pos, itemCount)
 
             // If removing first item, position can be -1; in which case, don't smoothScro
             if (pos == RecyclerView.NO_POSITION)
